combined projectContext
